fix(11-02): fix clippy warnings in pipeline/ modules
- Convert /// to //! inner doc comments to fix empty_line_after_doc_comments
- Add #[allow(clippy::too_many_arguments)] to spawn() and run_worker() in auto_store
- Fix redundant_pattern_matching in extraction/pipeline.rs (if let Err → is_err())
- Fix unnecessary_cast in gc/worker.rs (remove redundant as i64)
- Fix map_or → is_none_or in curation/algorithmic.rs
- Fix if_same_then_else in curation/worker.rs (remove duplicate branches)
- Fix wildcard_in_or_patterns in summarization/mod.rs and curation/mod.rs

+++ crates/memcp-core/src/pipeline/gc/worker.rs
@@ -1,4 +1,4 @@
-/// GC worker — implements salience-based pruning, TTL expiry, dry-run, and hard purge.
+//! GC worker — implements salience-based pruning, TTL expiry, dry-run, and hard purge.
 
 use anyhow::Result;
 
@@ -33,7 +33,7 @@ pub async fn run_gc(
     }
 
     // Step 3: Compute budget and fetch candidates
-    let prune_budget = (live_count - config.min_memory_floor as i64).max(0) as i64;
+    let prune_budget = (live_count - config.min_memory_floor as i64).max(0);
     let candidates = store
         .get_gc_candidates(config.salience_threshold, config.min_age_days, prune_budget)
         .await?;
